+++ RELEASE-NOTES.md
@@ -1,3 +1,4 @@
 ## 2.0
 
 * #1 Injectable SqlSessionFactory instance
+* #2 Allow registering TypeHandlers via MyBatisModuleExtender